chore: rename files and function from fulltext to search-api
