+++ src/api/search-api/all-entities.ts
@@ -1,0 +1,8 @@
+import { searchApiPost, searchApiGet } from './host';
+
+export const searchAllEntities = (body: any) => searchApiPost('', body);
+
+export const getRecentEntities = () => searchApiGet('/recent');
+
+export const extractEntities = (searchResponse: any) =>
+  searchResponse?.hits ?? [];
