--- src/api/search-fulltext-api/all-entities.ts
@@ -1,9 +1,0 @@
-import { searchFullTextApiPost, searchFullTextApiGet } from './host';
-
-export const searchAllEntities = (body: any) =>
-  searchFullTextApiPost('', body);
-
-export const getRecentEntities = () => searchFullTextApiGet('/recent');
-
-export const extractEntities = (searchResponse: any) =>
-  searchResponse?.hits ?? [];
