+++ src/pages/search-page/search-page-resolver.ts
@@ -1,7 +1,7 @@
 import memoize from 'lodash/memoize';
 import { resolve } from 'react-resolver';
 
-import { searchAllEntities } from '../../api/search-fulltext-api/all-entities';
+import { searchAllEntities } from '../../api/search-api/all-entities';
 import { parseSearchParams } from '../../lib/location-history-helper';
 import { PATHNAME_SEARCH } from '../../constants/constants';
 import { normalizeAggregations } from '../../lib/normalizeAggregations';
